refactor(Appointment): Componente individual para las operaciones de Cita
