+++ app.js
@@ -4,6 +4,7 @@ var express = require('express');
 var path = require('path');
 var cookieParser = require('cookie-parser');
 var logger = require('morgan');
+require("dotenv").config();
 
 const cors = require('cors');
 const mongoose = require('mongoose');
