+++ routes/index.js
@@ -3,6 +3,7 @@ const express = require('express');
 const router = express.Router();
 // Routers
 const patientRouter = require('./patientRouter');
+const appointmentRouter = require('./appointmentRouter')
 const userRouter = require('./userRouter');
 
 /* API Home page. */
@@ -13,5 +14,6 @@ router.get('/', (req, res) => {
 // Router linking
 router.use('/patients', patientRouter);
 router.use('/users', userRouter);
+router.use('/appointments', appointmentRouter)
 
 module.exports = router;
