+++ routes/patientRouter.js
@@ -19,10 +19,4 @@ router.put('/:doctorId/patient/:patientId', auth.authenticateToken, patientContr
 // Delete
 router.delete('/:doctorId/patient/:patientId', auth.authenticateToken, patientController.delete);
 
-// Add Appointment
-router.post('/:doctorId/create_appointment', auth.authenticateToken, patientController.createAppointment);
-
-// Get Appointments
-router.get('/:doctorId/appointments', auth.authenticateToken, patientController.appointments);
-
 module.exports = router;
